Change mechanism for component evaluation, support TLA vars
Previously components were evaluated by creating a megacomponent that imported
JSON, YAML and JSONNET files. In particular, jsonnet files were `imported`
which meant that they had no top-level.

We change this to evaluate every component in its own jsonnet VM such that the
top-level nature of the component jsonnet is maintained allowed for TLA vars to
be passed in.

Processing components like this leads to 30-40% overhead so we need to evaluate
components concurrently. This is done with a default concurrency of 5. The user
can change this number using the persistent `--eval-concurrency` flag.

This change has the following benefits and no downside (as far as I can tell)

* Allow for tla variables to be used in jsonnet components
* Reduce time to evaluate a large set of components
* Produce better ereror messages that have the failed component name when things
  do not load correctly.