Change mechanism for component evaluation (#6)
Previously components were evaluated by creating a megacomponent that imported
JSON, YAML and JSONNET files. This provided little to no context for errors.

We change this to evaluate every component in its own jsonnet VM.
Processing components like this sequentially  leads to 30-40% overhead so we need to evaluate
components concurrently. This is done with a default concurrency of 5. The user
can change this number using the persistent `--eval-concurrency` flag.

This change reduces the time to evaluate a large set of components _and_ produces better messages
that have the failed component name when things do not load correctly.