add experimental support to expand helm templates in qbec
add a new native function called `expandHelmTemplate` that accepts a chart name,
values as a JSON object and additional helm options. Since native functions do not
have caller context, the caller needs to pass a `thisFile` value in the options
object for relative references to charts succeed. This should be set to `std.thisFile`

This allows a jsonnet caller to produce a list of objects by running `helm template`
and return them, possibly modifying them as needed.