add structured support for external and top-level variables in qbec
Refactor the options structs and interfaces such that config is now
created in the commands package. This allows for a single implmentation
for real code and tests such that the newly introduced interactions
between VM config and app can be handled in one place and tested.

Add a `vars` section in qbec.yaml that allows definition of external
and top-level variables. External variables can have defaults that
are used when not specified on the command line. Top level variables
must be associated with components such that these are only set
for the components that need them.

Add `-V` and `-A` shorthands (similar to `jsonnet eval`) to set
string vars for external and top-level variables from the command line.
Code values are also supported but do not have short flags for
assigment.

Introduce a `--strict-vars` flag that restricts commands from
only setting up variables that are defined in qbec and requires
them to set _all_ such variables at the time of command invocation.