implement app-tag feature for scoped GC, fixes #12 (#13)
Add a CLI argument called `--app-tag` that allows the user to specify a tag scope for
a qbec command. When this is specified, GC is scoped to app ,env, and tag.

Expose additional jsonnet variables to expose tag name and default namespace.
Add an attribute in qbec YAML to autosuffix default namespace on tagged runs.