change diff implementation to show object content on adds and deletes
Previously, the diff showed a single line for adds and deletes (e.g.
'object not found on server' / 'object not found locally')

This is extremely annoying for adds where the user would typically prefer
to see the content. On the other hand, we may have made the delete case
more annoying by showing the full content of objects to be deleted.

It's still better to be consistent on both add and delete and we could
potentailly add a flag later to suppress full content of deletes.