add duplicate object checks and fail fast if duplicates found
currently qbec doesn't perform any duplicate checking which can
cause indeterministic behavior based on which version of the
object "wins" the race for backend updates.

This commit introduces a deep duplicate check for commands
that have access to cluster metadata and a shallow version
for the show command that typically doesn't need cluster metadata.

The deep check will detect duplicates like 2 deployments with the same
namespace and name that are marked, say, extensions/v1beta1 and
apps/v1 respectively.