improve shell integration: add env list|vars commands
Add 2 new commands `env list` and `env vars`.

The first prints a list of known environments. This is useful for
scripts that need to iterate over qbec environments.

The second prints environment variables in eval format that can be
use to run kubectl or other programs with a specific kubeconfig,
context, cluster and namespace that qbec would derive for a
specific environment.