initial implementation of wait after apply, fixes #35 (#43)
Add a `--wait` option to the `apply` command, which waits for readiness of all deployments, daemonsets and statefulsets similar to how it is implemented by `kubectl rollout status`.
Waiting is done up to a maximum time that can be controlled by the `--wait-timeout` flag. The default is 5 minutes.

Daemonsets and statefulsets must be created with the `RollingUpdate` strategy in order to be
considered for the wait.

The `--wait` flag is not on by default. A future release may change this behavior.

There are other objects for which we could legitimately wait on - for example: transient pods, job objects, load balancer services, persistent volume claims. The initial implementation has been kept simple to see how it works for folks before adding new features.