propagate --clean flag for postprocessor use
add a new ext var called `qbec.io/cleanMode` that is set to "on" or "off".
It is only ever set to "on" for the `show --clean` command. This can be
used by a post-processor to refrain from adding non-functional labels
and metadata to provide a "real clean" experience for the developer.