fix pristine object retrieval
The pristine reader is supposed to return a pristine object from qbec annotations.
When this is not found, it is either supposed to return a live object for diffs
or an empty object for patches. It was currently returning the live object for both
causing the patch to try and remove fields set by kubernetes. This is fixed by
correctly honoring the includeFallback flag.

Also, use kubectl annotations as a pristine source when qbec annotations are not found.
In this case also add the original k8s annotation into the deserialized object
because from qbec's viewpoint, this is a diff that needs to be reconciled. In other
words, by doing this, we will make qbec delete the original kubectl annotation
and "adopt" the object correctly.