sync up patch logic with kubectl for deviant case, fixes #51 (#53)
qbec tries a strategic merge patch when a schema is available for a type,
unlike kubectl which only does this when a schema is available _and_ the
type is registered in code (i.e. not a custom resource).

This commit fixes the patch logic to make it the same as on kubectl.
On Kubernetes 1.15 and above, this will actually make a difference since
custom types are returned in the schema.