Allow for component sub-directories at one level of nesting only
Currently, components are jsonnet, json or yaml files directly under the components/ directory.

Take a leaf out of static websites and allow a one-level subdirectory to be considered as a component if:

* it has an index.jsonnet file (this is the file that is loaded in this case)
* it has an index.yaml file in which case all yaml and json files in the directory are loaded as part of the component

This change is technically backwards-incompatible since there may be existing subdirectories for some projects that
may have one of the special files but in practice the impact of this should be minimal to non-existent.