allow custom qbec processing with qbec-specific annotations on objects
* namespace for such annotations is directives.qbec.io/
* apply-order: allows for a custom apply order (numeric)
* lazy-type: indicates that the type of an object may be lazily created by a workload that is also deployed
* update-policy: never indicates that qbec shouldn't update an object that exists
* delete-policy: never indicates that an object should not be deleted by qbec.

When an object is not deleted because of a delete policy, qbec ensures that its corresponding namespace will also
not be deleted.

The default and kube-system namespaces will also never be deleted independent of annotations.

The update and delete policies are consulted from the object in etcd metadata and not the local object. The other
annotations are consulted from the local object.