fix implementation of namespace override to affect all users of this attribute (#122) fixes #120 fixes #121
In the previous commit, overrides to the default namespace from the command line
only affected the namespace setup for the remote connection. In particular,
the namespace name exposed by qbec to the user was not updated with the forced value.

Change this such that the app object itself knows about the override and returns it
when set. This correctly affects any caller that may be using the namespace attribute.