add support for importing a glob of files with either import or importstr semantics (#153)
This change introduces the ability to import a bag of data or libsonnet files.
It provides the following import syntax:

    import 'glob-import:*.json'
        and
    import 'glob-importstr:*.yaml'

The first form looks for all files that match the wildcard and returns an object keyed by
relative file name with values importing the actual files. i.e. something like:

    {
        'a.json': import 'a.json',
        'b.json': import 'b.json',
    }

The second form is very similar except that it uses 'importstr' for inner imports. The returned
code looks like:

    {
        'a.yaml': importstr 'a.yaml',
        'b.yaml': importstr 'b.yaml',
    }

This allows for loading optional overlay files, user data files and such without having to write out
an import statement in the calling code for each file imported.

This importer is enabled by default in qbec.