add --wait-all in addition to --wait to wait on all objects
Currently, if --wait is specified and the apply times out, if apply is run
for a second time it will cheerfully succeed because it has already updated
the deployment and doesn't wait for it again.

Guard against this by allowing the caller to wait on all objects independent
of whether they were changed. In the current commit, the --wait-all flag
has a default of false to preserve backwards compatibility. The next minor
version will change the default of both the --wait and --wait-all flags to true