Lint (#240)
add lint functionality to qbec

Adds a "alpha lint" command to qbec that uses the jsonnet linter and supports qbec vm extensions like glob imports.
It does not yet support data sources completely. This will happen in a subsequent change.

The lint command typically requires a qbec app to be present. This can be turned off on the command line to provide the equivalent of jsonnet-lint.

Adds fail-fast and verbose options to both fmt and lint commands with sensible defaults.