add a native function to render YAML documents (#250)
expose a renderYaml function in the qbec VM that is compatible with qbec fmt. If it is passed an array, it will render multiple YAML documents, otherwise will render a single document. If you want an array to be serialized as a single document, create a one element array that has this value.