experimental: write component definitions in cue-lang
This is initial foray into what supporting cue with qbec might look like.

A component can be written with cue in addition to jsonnet, json and yaml. Implementation is bare-bones and is not production ready
Cue packages are supported but not well-tested for mutltiple cases.

Brainstorming needed  how this would surface in areas like envs, params, etc.